converted transport to trait and converted bytes to stdio transport struct
